+++ src/commands/restore.ts
@@ -1,4 +1,4 @@
-import { CommandContext } from './index';
+import { CommandContext } from './types';
 
 /**
  * rifler.restore - Restore minimized panel from status bar
